feat: Implement updating of priority of tasks
Tasks now have priority attributes but no way of changing it. The
default value is LOW but we want for users to be able to change it
to HIGH and back to LOW.

Add methods that change the priority based on user input in
Task.java and TaskList.java. Also, add the logic in
UpdatePriorityCommand.java to handle user input and call on the
method.

This ensures that all methods related to Tasks are abstracted away
inside of Task.java and TaskList.java. By using
UpdatePriorityCommand.java to handle user input also ensures that
no bad input goes to Task.java and TaskList.java as handling bad
input is not part of their jobs.