feat: Add filtering to task display
With new priority fields, tasks are able to have LOW and HIGH
priorities. User is able to change them but they are not yet able
to see only LOW priority tasks or only HIGH priority tasks.

Users should be able to see only LOW or HIGH priority tasks to
check up on the more urgent tasks, housekeeping or more.

Add new command filter which takes 1 argument: HIGH or LOW. Then,
check which tasks fulfills the filter and hide those that are not.

This is similar to the approach taken in the find command. In
further versions, the combining of these 2 commands could be
explored.